re-enable export IW tests
